Redact API key from debug logging
This redacts the API key, which ends up in Authorization header in debug
logs. This helps both reduce the size and also protect secrets.

As a side effect, this adds a mockwebserver test, matching version with
the okhttp client used in openapi generated code. This part is important
as we improve the debug experience for watch events later.

Signed-off-by: Adrian Cole <adrian@tetrate.io>