AuthenticationRefresher wrapper solution
This commit introduces the AuthenticationRefresher as an implementation
of Authentication. It's purpose is to warp another Authentication
instance and refresh it very n seconds by calling `provide` on the
wrapped Authentication object.

This code is intedned to be used a solution to issue #2438.

Any Authentication object that provides credentials, that must be
refreshed, to the ApiClient do not have a way to do so.

There are two ways to use the AuthenticationRefresher with
ClientBuilder. The first option is to use the
ClientBuilder.setAuthentication(...) method. An Authentication instance
can be wrapped in an AuthenticationRefresher and passed into the
setAuthentication method. For example,

```
ClientBuilder.standard().setAuthentication(
  //wrap an auth and refresh it every 15 minutes (900s)
  new Authentication(someDelegateAuthenticationInstace, 900)
);
```

This is integrated up into the ClientBuilder.build() method by adding a
new authenticationRefreshSeconds field and getter/setter pair. If a
refresh interval is set, the Authentication object used by ClientBuilder
will be wrapped with an AuthenticationRefresher.