Updates to the AuthenticationRefresher timer
The AuthenticationRefreher will now keep a instance-level reference to
the current Executor. A singleton style synchronization ensures there is
only ever one Executor active.

Additionally, the APIClient passed into .provide is now heald as a
WeakReference by the Executor. if the APIClient is ever garbage
collected the Executor is cancelled and nulled.