Add params to publish-gen-image.sh
This change makes it convenient to publish a copy of crd-model-gen to a private container registry.

Refactor client-java-contrib/publish-gen-image.sh to accept parameters via environment variables. At the same time, refactor scripts/publish-crd-model-gen-image.sh to remove duplicated logic. This also serves as an example of how to invoke publish-gen-image.sh.

In case there are existing workflows that depend on these two scripts, I have not renamed either of them, and they have the same behavior when no environment variables are provided.